requirements.txt
torch
torchvision
pudb
pytorch-lightning
einops
regex
ftfy
cython
WebDataset
pillow
wandb
sklearn
tensorboard
mkl-service